Version bump and namespace changes
