+++ src/ZPLForge.XmlSerialization/LabelXmlSerializer.cs
@@ -2,7 +2,6 @@
 using System.IO;
 using System.Xml.Serialization;
 using ZPLForge.Contracts;
-using ZPLForge.XmlSerialization.Helpers;
 
 namespace ZPLForge.XmlSerialization
 {
