+++ src/ZPLForge.XmlSerialization/SimpleObjectMapper.cs
@@ -6,7 +6,7 @@ using System.Linq;
 using System.Linq.Expressions;
 using System.Reflection;
 
-namespace ZPLForge.XmlSerialization.Helpers
+namespace ZPLForge.XmlSerialization
 {
     internal static class SimpleObjectMapper
     {
