+++ src/ZPLForge.XmlSerialization/ZPLForge.XmlSerialization.csproj
@@ -4,7 +4,7 @@
     <TargetFramework>netstandard2.0</TargetFramework>
     <PackageRequireLicenseAcceptance>false</PackageRequireLicenseAcceptance>
     <GeneratePackageOnBuild>true</GeneratePackageOnBuild>
-    <Version>1.0.0-beta</Version>
+    <Version>1.0.0</Version>
     <Authors>Christopher Hanzlik</Authors>
     <Product>ZPLForge.XmlSerialization</Product>
     <Copyright>Copyright (c) 2020 Christopher Hanzlik</Copyright>
