+++ src/ZPLForge/Builders/EllipseBuilder.cs
@@ -1,5 +1,4 @@
-﻿using ZPLForge.Builders.Abstractions;
-using ZPLForge.Common;
+﻿using ZPLForge.Common;
 using ZPLForge.Configuration;
 
 namespace ZPLForge.Builders
