+++ src/ZPLForge/Builders/LabelBuilder.Elements.cs
@@ -1,5 +1,4 @@
 ﻿using System;
-using ZPLForge.Builders.Abstractions;
 
 namespace ZPLForge.Builders
 {
