+++ src/ZPLForge/Presets/BlackMarkSensingPreset.cs
@@ -1,19 +1,23 @@
 ﻿using System;
+using System.Collections.Generic;
+using System.Text;
 using ZPLForge.Common;
 using ZPLForge.Contracts;
 
-namespace ZPLForge.Builders.Presets
+namespace ZPLForge.Presets
 {
-    public class WebSensingPreset : ILabelPreset
+    public class BlackMarkSensingPreset : ILabelPreset
     {
-        public WebSensingPreset(int printWidth, PrintMode? printMode, MediaType? mediaType)
+        public BlackMarkSensingPreset(int printWidth, int blackMarkOffset, PrintMode? printMode, MediaType? mediaType)
         {
             PrintWidth = printWidth;
+            BlackMarkOffset = blackMarkOffset;
             PrintMode = printMode;
             MediaType = mediaType;
         }
 
         public int PrintWidth { get; }
+        public int BlackMarkOffset { get; set; }
         public PrintMode? PrintMode { get; }
         public MediaType? MediaType { get; }
 
@@ -25,7 +29,8 @@ namespace ZPLForge.Builders.Presets
             label.PrintWidth = PrintWidth;
             label.PrintMode = PrintMode;
             label.MediaType = MediaType;
-            label.MediaTracking = MediaTracking.WebSensing;
+            label.BlackMarkOffset = BlackMarkOffset;
+            label.MediaTracking = MediaTracking.MarkSensing;
         }
     }
 }
