+++ src/ZPLForge/Presets/ILabelPreset.cs
@@ -1,6 +1,6 @@
 ﻿using ZPLForge.Contracts;
 
-namespace ZPLForge.Builders.Presets
+namespace ZPLForge.Presets
 {
     public interface ILabelPreset
     {
